Bugfix: Fix integer casting of learning rate
- Modified argument list for 'run' and 'gradDesc' methods
- Changed type of learning rate from int to double
- Prevent low learning rate values rounding down to zero

+++ include/Network.h
@@ -24,13 +24,13 @@ private:
 public:
     Network(const LayerData& layerData, Eigen::MatrixXf X, Eigen::MatrixXf Y);
 
-    Eigen::MatrixXf run(int numIterations, int learningRate);
+    Eigen::MatrixXf run(int numIterations, double learningRate);
 
     Eigen::MatrixXf forwardProp();
 
     void backProp();
 
-    void gradDesc(int learningRate);
+    void gradDesc(double learningRate);
 };
 
 #endif
